+++ tests/testthat/test-serialization.R
@@ -77,3 +77,14 @@ test_that("data.frames with many columns survive roundtrip", {
 
   ensure_round_trip(sc, data)
 })
+
+test_that("data.frames with many columns don't cause Java StackOverflows", {
+  skip_on_cran() && skip_unless_verbose("Skipping slow serialization test")
+
+  n <- 5000
+  df <- matrix(0, ncol = 5000, nrow = 2) %>% as_data_frame()
+  sdf <- copy_to(sc, df, overwrite = TRUE)
+
+  # the above failed with a Java StackOverflow with older versions of sparklyr
+  expect_true(TRUE, info = "no Java StackOverflow on copy of large dataset")
+})